+++ __init__.py
@@ -87,7 +87,7 @@ utils_modules = [
     "cad_module", "sv_bmesh_utils", "sv_viewer_utils", "sv_curve_utils",
     "voronoi", "sv_script", "sv_itertools", "script_importhelper",
     "csg_core", "csg_geom", "geom", "sv_easing_functions",
-    "snlite_utils", "snlite_importhelper", "context_managers",
+    "snlite_utils", "snlite_importhelper", "context_managers", "sv_bgl_lib",
     # UI text editor ui
     "text_editor_submenu", "text_editor_plugins",
     # UI operators and tools
